Fix deployment if git checkout becomes unclean
The checkout might be unclean if a script changes and fails to parse old
pack.yaml files. This happens, for example, with the test pack. The test
pack moved from version 0.3 to version 0.3.0 at when semantic versioning
was introduced/enforced. But 0.3 is not a valid semantic version, so one
of the scripts ends up collapsing with the old version number. Though we
just continue past the bad tag, the git checkout is no longer clean so
it fails to finish the rest of the deployment script.